requirements.txt
selenium
python-dotenv
webdriver-manager
google-generativeai==0.8.6